refactor: Change the structure of the project
