--- backend/django/templates/django/login.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <title>42 OAuth 로그인</title>
-</head>
-<body>
-    <button id="loginBtn">42로 로그인</button>
-
-    <script>
-        document.getElementById('loginBtn').addEventListener('click', function() {
-            window.location.href = '/api/login/';
-        });
-    </script>
-</body>
-</html>
