+++ backend/login/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class AnsankingConfig(AppConfig):
+class LoginConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'ansanking'
+    name = 'login'
