+++ backend/requirements.txt
@@ -13,6 +13,4 @@ setuptools==71.1.0
 sqlparse==0.5.1
 typing_extensions==4.12.2
 uritemplate==4.1.1
-
-# feature01-42OAuth_Login 에서 추가됨
 requests==2.31.0
