+++ frontend/src/components/Friends-Info.js
