+++ frontend/src/components/Friends.js
