+++ frontend/src/core/Component.js
