+++ run_frontend.sh
@@ -2,7 +2,7 @@
 
 # TEST: 개발 시 로컬 실행용
 
-cd frontend/vite
+cd frontend
 
 npm install
